Breakout widgets into separate modules. Auto-import current modules at pymeasure/display/widgets/__init__.py for backwards compatibility
