+++ pymeasure/display/widgets/results_dialog.py
@@ -1,0 +1,105 @@
+#
+# This file is part of the PyMeasure package.
+#
+# Copyright (c) 2013-2022 PyMeasure Developers
+#
+# Permission is hereby granted, free of charge, to any person obtaining a copy
+# of this software and associated documentation files (the "Software"), to deal
+# in the Software without restriction, including without limitation the rights
+# to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+# copies of the Software, and to permit persons to whom the Software is
+# furnished to do so, subject to the following conditions:
+#
+# The above copyright notice and this permission notice shall be included in
+# all copies or substantial portions of the Software.
+#
+# THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+# IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+# FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+# AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+# LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+# OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
+# THE SOFTWARE.
+#
+
+import logging
+
+import os
+import pyqtgraph as pg
+
+from ..curves import ResultsCurve
+from ..Qt import QtCore, QtGui
+from ...experiment.results import Results
+from .plot_widget import PlotWidget
+
+log = logging.getLogger(__name__)
+log.addHandler(logging.NullHandler())
+
+
+class ResultsDialog(QtGui.QFileDialog):
+    def __init__(self, columns, x_axis=None, y_axis=None, parent=None):
+        super().__init__(parent)
+        self.columns = columns
+        self.x_axis, self.y_axis = x_axis, y_axis
+        self.setOption(QtGui.QFileDialog.DontUseNativeDialog, True)
+        self._setup_ui()
+
+    def _setup_ui(self):
+        preview_tab = QtGui.QTabWidget()
+        vbox = QtGui.QVBoxLayout()
+        param_vbox = QtGui.QVBoxLayout()
+        vbox_widget = QtGui.QWidget()
+        param_vbox_widget = QtGui.QWidget()
+
+        self.plot_widget = PlotWidget("Results", self.columns,
+                                      self.x_axis, self.y_axis, parent=self)
+        self.plot = self.plot_widget.plot
+        self.preview_param = QtGui.QTreeWidget()
+        param_header = QtGui.QTreeWidgetItem(["Name", "Value"])
+        self.preview_param.setHeaderItem(param_header)
+        self.preview_param.setColumnWidth(0, 150)
+        self.preview_param.setAlternatingRowColors(True)
+
+        vbox.addWidget(self.plot_widget)
+        param_vbox.addWidget(self.preview_param)
+        vbox_widget.setLayout(vbox)
+        param_vbox_widget.setLayout(param_vbox)
+        preview_tab.addTab(vbox_widget, "Plot Preview")
+        preview_tab.addTab(param_vbox_widget, "Run Parameters")
+        self.layout().addWidget(preview_tab, 0, 5, 4, 1)
+        self.layout().setColumnStretch(5, 1)
+        self.setMinimumSize(900, 500)
+        self.resize(900, 500)
+
+        self.setFileMode(QtGui.QFileDialog.ExistingFiles)
+        self.currentChanged.connect(self.update_plot)
+
+    def update_plot(self, filename):
+        self.plot.clear()
+        if not os.path.isdir(filename) and filename != '':
+            try:
+                results = Results.load(str(filename))
+            except ValueError:
+                return
+            except Exception as e:
+                raise e
+
+            curve = ResultsCurve(results,
+                                 x=self.plot_widget.plot_frame.x_axis,
+                                 y=self.plot_widget.plot_frame.y_axis,
+                                 # The pyqtgraph pen width was changed to 1 (originally: 1.75) to
+                                 # circumvent plotting slowdown. Once the issue
+                                 # (https://github.com/pyqtgraph/pyqtgraph/issues/533) is resolved
+                                 # it can be reverted
+                                 pen=pg.mkPen(color=(255, 0, 0), width=1),
+                                 antialias=True
+                                 )
+            curve.update_data()
+
+            self.plot.addItem(curve)
+
+            self.preview_param.clear()
+            for key, param in results.procedure.parameter_objects().items():
+                new_item = QtGui.QTreeWidgetItem([param.name, str(param)])
+                self.preview_param.addTopLevelItem(new_item)
+            self.preview_param.sortItems(0, QtCore.Qt.AscendingOrder)
